Refactor: change file names
The file names were changed to follow the Dart convention, which states that file names should be in camel_snake_case.
See https://dart.dev/effective-dart/style
