+++ lib/app/widgets/atenea_button.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import '../values/AppTheme.dart';
+import '../values/app_theme.dart';
 
 class AteneaButton extends StatelessWidget {
   final String? text;
   final VoidCallback onPressed;
-  final Color backgroundColor;   
+  final Color backgroundColor;
   final TextStyle? textStyle;
   final String? svgIcon;
 
@@ -15,16 +15,8 @@ class AteneaButton extends StatelessWidget {
 
   final bool enabledBorder;
 
-  AteneaButton({
-    super.key, 
-    this.text,
-    required this.onPressed,
-    this.backgroundColor = AppColors.primaryColor, 
-    TextStyle? textStyle,
-    this.enabledBorder = false,
-    this.svgIcon
-  }) : textStyle = textStyle ?? AppTextStyles.builder(color: AppColors.ateneaWhite, size: FontSizes.h4); // Lo hacemos aquí.
-
+  AteneaButton({super.key, this.text, required this.onPressed, this.backgroundColor = AppColors.primaryColor, TextStyle? textStyle, this.enabledBorder = false, this.svgIcon})
+      : textStyle = textStyle ?? AppTextStyles.builder(color: AppColors.ateneaWhite, size: FontSizes.h4); // Lo hacemos aquí.
 
   @override
   Widget build(BuildContext context) {
@@ -35,26 +27,24 @@ class AteneaButton extends StatelessWidget {
         shape: RoundedRectangleBorder(
           borderRadius: BorderRadius.circular(borderRadius),
         ),
-        side: enabledBorder  // Condicional para mostrar borde solo si `hasBorder` es `true`
-              ? const BorderSide(
-                  color: AppColors.primaryColor,  // Color del borde
-                  width: 1.5,  // Grosor del borde
-                  style: BorderStyle.solid,  // Estilo sólido
-                )
-              : BorderSide.none, 
+        side: enabledBorder // Condicional para mostrar borde solo si `hasBorder` es `true`
+            ? const BorderSide(
+                color: AppColors.primaryColor, // Color del borde
+                width: 1.5, // Grosor del borde
+                style: BorderStyle.solid, // Estilo sólido
+              )
+            : BorderSide.none,
       ),
       onPressed: onPressed,
       child: Row(
         mainAxisSize: MainAxisSize.max,
         mainAxisAlignment: MainAxisAlignment.center,
         children: [
-
-          if (text != null) 
-          Text(
-            text!,
-            style: textStyle,
-          ),
-
+          if (text != null)
+            Text(
+              text!,
+              style: textStyle,
+            ),
           if (svgIcon != null)
             Padding(
               padding: EdgeInsets.only(left: text != null ? 8.0 : 0.0),
