+++ lib/domain/repositories/session_repository.dart
